backend/requirements.txt
fastapi
uvicorn
pyyaml
pytest
httpx
beautifulsoup4
